requirements.txt
absl-py==0.7.0
altgraph==0.16.1
astor==0.7.1
certifi==2018.11.29
cloudpickle==0.7.0
dask==1.1.1
decorator==4.3.2
future==0.17.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
libtiff==0.4.2
lxml==4.3.1
macholib==1.11
mahotas==1.4.5
Markdown==3.0.1
mkl-fft==1.0.10
mkl-random==1.0.2
networkx==2.2
numpy==1.15.4
numpy-stl==2.9.0
opencv-python==4.0.0.21
pefile==2018.8.8
Pillow==5.4.1
protobuf==3.6.1
PyInstaller==3.5.dev0+73dfb3add
pypiwin32==223
pypng==0.0.19
PyQt5-sip==4.19.14
python-utils==2.3.0
PyWavelets==1.0.1
pywin32==224
pywin32-ctypes==0.2.0
scikit-image==0.14.2
scipy==1.2.0
sip==4.19.8
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
toolz==0.9.0
tornado==5.1.1
Werkzeug==0.14.1
wincertstore==0.2

